[TECH-11069] Load products and stock in the menu
In order to load products we need to have access to the cutoff date. So we had to create cutoff date as a helper in order to use it. We take the date from the basket (state) and we use that to calculate the cutoff date. For now, we use a constant to subtract from the date as cutoff always happens 3 days before the delivery.