Fix bug with product modal being cut off
Previously, there was a bug which caused the product modal to be cut off if initially loaded on a small screen which was later resized to be larger. This was due to the getFullHeight function being used to calculate a height which was then to be used in a inline style, however it was not recalculated when the screen was resized. Use of this function has been replaced with 100vh.