TECH-11359 Replace surcharge prop with surchagePerPortion
Previously, the surcharge prop was passed down to the Button component through many other components higher in the tree. This prop drilling makes the flow of data difficult to trace, and is here unnecessary as none of the inbetween components actually make use of the surcharge prop, they simply pass it on. This commit improves this situation in two ways:
- The ButtonContainer takes the relevant surcharge from the store and passes directly to Button
- Rather than consuming the surcharge prop and then converting this to surchargePerPortion within the Button component, the ButtonContainer now handles this logic and passes the surchargePerPortion only