TECH-11359 Migrate tests for Buttons component from Mocha to Jest
The previous Mocha tests for the Buttons component have been manually migrated to the Jest __tests__ directly so that they run as part of the CI process. Some existing tests have been enhanced, and new tests have been added. The old test file has been deleted.

Shallow rendering has been used for the 'appearance' in order to make the tests more easily understandable. For example, when checking for the presence of 3 Segments in the 'clicked' button state, if the component was mounted then the test would actually have to check for the presence of 4 segments due to the fact that the Buutton component (rendered by Buttons component) contains a segment. Whilst there are workarounds for this, they make the tests much more difficult to understand just by reading the descriptions.