TECH-11359 Style buttons to handle wrapping and disappearance of surcharge
The buttons have been styled so as to match the behaviour of the Optimizely experiment GOU 001. On Buttons have that have not been clicked i.e. the recipe has not been selected, the Surcharge component wraps under the ADD RECIPE text on smaller screen sizes, whereas for Buttons which have been clicked, i.e. the recipe has been selected, the Surcharge component disappears on smaller screen sizes. To achieve this here, the Surcharge components have been wrapped in divs which have the appropriate styles applied so as to not affect the Surcharge component styles in all cases.

Because much of the styling is only required on Buttons which contain a Surcharge component and are displayed only in the Grid view (not when viewing a recipe in detail), a helper function getSurchargeGridClass has been defined as a class method on the Buttons component, which applies the provided className to an element only if there is a surcharge and the user is not viewing in detail. Tests have been added to ensure that this helper function is working correctly.