[TECH-11409] Refactor: Create new selector using getOkRecipeIds
Previously, multiple component containers computed a list of unavailable recipes, using the okRecipeIds list to do so.  Given the new getOkRecipeIds selector from the previous commit, this commit has created a new selector to get the list of unavailable recipes - getUnavailableRecipeIds - which can be reused across the multiple components, and again has the performance benefits that come with memoization. Furthermore, this commit has simplified the Details component by passing the list of unavailable recipes directly as a prop, rather than having to compute it within the render method.

Use new selector for box summary components