[TECH-11409] Refactor: Create selector to compute okRecipeIds
Currently, the Details component computes a list of available recipes - okRecipeIds -  in its render method, by calling the okRecipes function. This okRecipes function is an expensive calculation, and therefore call this every time the component is rendered is not good for performance. This commit has  introduced a getOkRecipeIds selector, which has been built using reselect and is therefore memoized. This memoiziation is good for performance, especially considering that the selector can be reused in other container components (e.g. BoxSummaryDesktop - as is done in a later commit).