[TECH-11409] Create BoxProgressAlert component and implement in Details component
A BoxProgressAlert component has been created, which is responsible for rendering the relevant Alert (correct color and correct message depending on how many recipes are selecred). This has been implemented in the existing Details component. The new Alert colors have been added into colors.css so that they are available to the component.