[TECH-11864] Fix selenium-download version
Previously, a later version of selenium-download was being installed, but this meant that meant that card expiry and CVV were not being entered during tests, causing the tests to fail. Version 2.0.7 has proven to work with these tests. The exact reason is unknown, but possibly a ersion mismatch between the selenium standalone server and chromedrive (installed by this selenium-download package) and the nightwatch browser.keys() syntax being used for card expiry and CVV fields.