Add underline on hover to ribbon text when feature flag is true
When the food brand page feature flag is true we want to show the ribbons on the recipe cards are clickable by adding in an underline and hover. This was built in css using before elements so that on the hover the text of the ribbon does not also get a darker shade, but the background color does. This is the same for the hover state and the focus state so it is accesible and the blue border is removed so it is just the underline and color change that is visible.

The underline should only appear when the feature flag is true and you can click through on the range bange. Then when you tab through using your keyboard instead of the blue outline default showing we have created hover and focus states add an opacity of 0.1 over the different components that make up the range badge.

Moves the uppercase to the css file rather than in the js file

The Tests:
Reflects uppercase change in tests and class names

Update test to match style updates

We moved the uppsercase in rangebadge to the css and so this wasn't reflected in the tests