Improves sticky navbar behaviour
We removed all of the check scroll methodology and then the changing of classes at certain scroll depths. Replaced this with a simple position: sticky and associated css. We moved the sticky bar container to the top of the section so that it is sticky relative to it's parent, the page, so it doesn't scroll of page.  The border was created as a new separate div at the bottom of the container that has a top of 59px - this will be updated in the next commit to be responsive to the height of the above banner.