Element height util using querySelector
Created a util to help us find the height of an element based on its selector

We needed to also pass document as a param into getElementHeight so we could write a good test. Have updated this in the foodbrand page and then in the util. The test checks for a mock doc to return the correct value and then checks that it calls query selector with the correct selector