Style border based on height of sibling element
Created logic and pulled in the new height util to tell the border div when to become sticky as the page scrolls. The height of the title container changes height based on the device, so this now allows the logic to work this out and apply the style direct to the border div.

Changed test to find the div with class .border and created a mock function to call the new utils getElementHeight and pass back a 50px height to use in the styling of the border element