Adds food brand page tracking to basket add and remove
Used the selector to find the food brand data from redux and then pass the slug into the tracking data. When recipe added or removed and not on a food brand page then shows 'false' and when on a food brand page it will show the slug of that food brand.