Header test for phone number
Previously there was a skipped test (that failed) checking to see if the phone number was visible in the Header. This test was trying to use mount which complicated the test - the Header component renders Redux-connected child components. A new shallow test has been written to test for phone number, and is in keeping with other tests in the file.