Add in selectCurrentCollection when component updates
We added this in as there was a bug where the collection Id was persisting if you went from menu to other pages and back to menu. The collection nav would say 'All Recipes' but would display the recipes from the last collection the user had clicked on.

Now we check to see if the previous collection id was different from the current one (usually all recipes) and if so then we reselect the current collection