[TECH-12223] Refactor tests for OrderConfirmation
Existing tests were confusing in terms of structure and they had a lot of repeated information. This commit is refactoring the structure of the tests, rewriting some of them, grouping some of them under the same describes and abstracting the rendering of the component. This should make them easier to read. Also this commit is dealing with the warnings thrown by the tests.

I also had to remove the `isRequired` from the products prop as the tests were covering the case when no products are being passed.

Co-authored-by: Gabriel Pasparan <gabriel.pasparan@gousto.co.uk>