[TECH-XXXX/zendesk-fix-error-handling] Fix error handling when zendesk instance fails
In some cases, like on mobile viewport, we have a rule set on GTM, where it doesn't load zendesk library. In our application, we were handling when the instance is not ready, but instead, we were throwing an error. I put a fix to use logger.notice and removed the Error.