[TECH-10811] Replace phone number regex generator with faker library in e2e
The Regexp library was generating a bad phone number and although the tests weren't failing now, they will with the changes that are expected in core and customer service, as we're trying to make the phone number validation more strict. We replaced it with faker, to follow a pattern that's already in other tests.

Co-authored-by: Chris Vibert <35846795+cpv123@users.noreply.github.com>