[TECH-12753] Display phone number validation error message on signup
Previously, if a customer could not sign up due to an invalid phone number, the message displayed was a generic error message that didn't specify the phone number as being the issue. Now, a phone number specific message is shown if the signup fails due to an invalid phone number.