[TECH-12795] Create routing selector to see if on My Gousto page
Unfortunately, creating reselect selectors that take generic arguments (e.g. pageName) is not recommended and so a selector has been created specifically to tell if you are on the My Gousto page.

The selector has been built to handle two types of store shape - the one that you would see in production (generateRoutingStore) and one that you would see when developing locally (generateAlternateRoutingStore).

Note that the composed selector (locationAtMyGousto) has not been tested as a pure unit test (reselect offers a .resultFunc method which allows you to mock the result of the input selectors and then test only the composed selector logic). This has not been done here due to the complexity of multiple store shapes.