[TECH-12712] Extract view for the Progress bar in menu in separate component
We have to render that component in two contexts now, depending on the user's authentication status. Initially, I extracted it in the same file and reused it, but it was looking messy. So I've decided to extract it to a separate component altogether, which helped with the separation of concerns a bit more, as it allowed us to move some props closer to where they are needed.