[TECH-12928] Create util function to generate categories from products
Utility function to generate a list of all categories (including product count for each) to display in the Product filter Nav Bar. This differs from the existing getCategories class method on Market.logic because it returns:
* Categories as an object
* A count of the number of products for each category
and therefore the categoriesForNavBar prop will change shape in following commits when this function is used.