[TECH-12928] Generate categories in container component using new util function
Instead of generating the categories from products in the Market.logic component, a selector has been used to generate this in the container and pass down as a prop. This allows us to make use of selector memoization, simplify the Market.logic component, and test the selector/utility function in more isolation.

This also means that categories are now an object, and that each category includes a count of products within in.