[TECH-12928] Replace Navbar and Dropdown component with ProductNavBar
A new ProductNavBar component has been build from two ui-components: CollectionsNavigation and CollectionsNavigationItem. Given that this new component handles screen sizes itself, it has replaced two existing component - Narbar for desktop, and Dropdown for mobile. Currently, the categories being passed as a prop do not include a count of products and so only the category title is shown - the count will come in a later commit.