[TC-935] Remove fetching of user orders in the header when on My Gousto page
This change was introduced in https://github.com/Gousto/gousto-webclient/pull/1104 in order to enable an Optimizely experiment. This is no longer needed because the My Gousto page has been migrated to the new stack.

This is not a straight git revert because there have been refactoring of some files and some of the changes in the initial commits are still valid.