[TC-928] Implement CardWithLink components in My Gousto header
The CardWithLink components have been implemented, but the content of the cards has been left as it currently is i.e. it does not match the designs exactly at this stage.

Exisiting tests have been updated so that they pass, but at this stage new tests have not been created.