[TC-928] Update order details copies to match new design
To show the order details as per the design, the functions in Header.logic that generate these copies have been updated. This in turn means that Header.presentation receives slightly different propTypes (an object rather than string for nextOrderMessage prop) and so the Header.presentation has been adapted to handle these and display the copies accordingly. Note that styling will come in the next commit.