[TC-981] Create selectors to order products with out of stock at the bottom
Selectors created allowing products to be returned with out of stock items at the end.

The getProductsWithOutOfStockLast selector includes converting the products to a JS object. This is required because the component receiving these products (Market) expects them as a plain JS object.

To test the order of products, the object keys as a string has been compared because jest's toEqual doesn't care about key/value order i.e. { a: 1, b: 2 } === { b: 2, a: 1 }