[TC-972] Pass the action_order query parameter coming from the menu to the orderConfirmationRedirect action
Before we had Menu passing this query parameter to oreder-confirmation page. But now we go to order-add-ons page, so we did passed this query parameter to the
order-add-ons page so it doesn't get lost in a previous commit. Here we are just grabbing that parameter from the url in order-add-ons page and pass it to
order-confirmation page