[TC-997] Move eligibity logic from GetHelp into Ingredients
We no longer want to call the validate endpoint on the top-level component. The idea was to move the validation logic into the section that makes sense, and, let the other pages free to be accessible by the customer.

This work was basically moving one logic and tests to another component with some very small refactoring.