Tc 1158/show gethelp tooltip yesterday (#1753)
* [TC-1158] Clean MyGousto Header tests a bit

A couple mock functions were passed as props to the component when mounting it all over the place. This is not needed as they are
not required and they default to () => {}. Now it only passes the mocks in the cases were they are tested

* [TC-1158] Add a tooltip that promotes the SSR link for the past delivery, when /show-tooltip endpoint says so

* [TC-1158] Zero eslint warnings in My Gousto Header

* [TC-1158] Remove unused util function

* [TC-1158] DRY the check to know if the tooltip should be shown

* [TC-1158] Move a couple of tests to the correct describe scope

In the previous ticket (TC-1154) of the epic to show the tooltip, this 2 tests were set in the wrong describe ('next order is for today' instead of 'next order'). It is true that at the moment the rule is when
'next order is for today' but that could change, and the FE logic should not know about that. It is the BE who tell us whether we should display it.

* [TC-1158]/CR Rename variables