Create a util function for testing async lifecycle methods: flushProm… (#1762)
* Create a util function for testing async lifecycle methods: flushPromises

When a lifecycle method like componentDidUpdate contains more than one promise running, something like this is needed to really test things after the lifecycle method has finished. At least if we use enzyme.
There is an interesting discussion here: https://github.com/airbnb/enzyme/issues/1587

* Use nextTick in flushPromises and move it to the _testing folder

Using nextTick instead of setImmediate it makes sure that the resolve of the promise will
be the next thing to be executed. setImmediate would place it the first in the queue but
it does not guarantee that nothing else will be place at the beginning of the queue in
the meanwhile.

It has been moved to _testing to make more obvious that it must not be used for browser
code, only for tests.