Tc 1182/restyle login (#1794)
* [TC-1182] align title to left

* [TC-1182] Change style and copy of the login button

* [TC-1182] Change "Remember Me" and "Forgot your password" styles

They needed to be both 16px, but we cannot change the size of <CheckBox> label because <CheckBox> can be used in other places. So a new property has been introduced for textSize.
They needed to be in one line when the modal width permits it, it is done with flex. Some margins where needed to make it break before they touch and give them some vertical separation when they break.

* [TC-1182] Use labels instead of placeholders

* [TC-1182] Narrow the max width of the Login modal

But not the other modals!
I had to use !important to override another !imporatnt

* [TC-1182] Fix easy wins eslint warnings for the files I touched

* [TC-1182]/CR extract classNames logic to make the render more readable

* [TC-1182]/CR Add isNarrow prop to <ModalPanel> instead of passing a custom className

* [TC-1182] FIX eslint warnings of the file I touched

And improve accessibility in the process!

* [TC-1182] Fix the tests I broke when fixing the defaultProps eslint warnings

* [TC-1182] Fix onSubmit function for Login

An accidental conflict resolution reverted the change to what it was before the recent refactor. This is restoring the right behaviour.

* [TC-1182] Move Forget password link in line with remember me on Login modal

This was the intended implementation but a bad conflict resolution broke that.

* [TC-1182] Simplify render logic for overlay in ModalPanel

The logic used a ternary with a null value as default when the same thing can be achieved simpler by negating the flag.

Co-authored-by: Razvan Spatariu <razvanspatariu@gmail.com>