[TC-1288] Fix 50 recipes appearing in the GetHelp wizard (#2022)
When coming via /get-help/check-eligibility journey you componentDidUpdate was necessary to load the recipes. However, in this case the eligible order is not fetched from an orderID, but extracted
from the 10 orders that were fetched already. When doing this, though, the recipe IDs of the order were not being taken, resulting on a fetch of recipes without filtering by id, which was bringing
50 recipes.