[TG-1888] Add app promotion banner to My Gousto (#2590)
* [TG-1888] Create AppAwarenessBanner

Also implement it behind the feature flag and when it's desktop viewport

* [TG-1888] Create isAppAwarenessEnabled feature

* [TG-1888] Add content to AppAwarenessBanner

* [TG-1888] Add flag to not render LimitedCapacityNotice when showAppAwareness is set to true

* [TG-1888] Bring user phone number to the banner

* [TG-1888] Create and add <InputWithButton /> to My Gousto

Wrap InputField + Button and apply some CSS modifications to them

* [TG-1888] Update ui-components

So it gets the latest updates for the InputField

* [TG-1888] Remove duplicate css height value

* [TG-1888] Change Gousto app image to a higher quality one

Also changes the banner body copy to the new fonts and colour

* [TG-1888] Fix broken tests

Both InputWithButton and MyGousto had a broken test. As I copied InputWithButton over from other repository, I forgot to update the import paths on the test file. And MyGousto renders the app awareness banner which has a container but was never `find` as it needed `connect` wrapped to its name

* [TG-1888] Adjust inputWithButton padding-top

So it's consistent to the others

* [TG-1888] Add react-transition-group to the package.json

* [TG-1888] Add test coverage for AppAwarenessBanner