[TG-2274] on clicking Build my box, redirect to wizard or to menu (#2727)
* [TG-2274] on clicking Build my box, send tracking event and redirect

* [TG-2274] add the space back that was removed accidentally

* [TG-2274] add an id to initialize a/b test in optimizely

* Revert "[TG-2274] add an id to initialize a/b test in optimizely"

This reverts commit 2dfa210c7054d79e8476d643c4893b5b8fce1fed.

We're initializing the experiment on waitUntil(loadFeatures is defined),
therefore we don't need a special element id.

* [TG-2274] update test snapshot

* [TG-2304] change wording on the Box Prices page

This is changed for both control and variation, i.e. it's a functional change
not related to the experiment.  Since it's so tiny, it's included in the PR
made for the experiment that touches this page.

* [TG-2274] use rem instead of px

The value corresponds to 10px because that's the value in
devbox-platform/code/gousto-webclient/src/src/components/LinkButton/LinkButton.css

which is included in the baseline, and we'd like to preserve the designed look
and feel of the Box Prices page for the variation.

* [TG-2274] add a unit test for the utility method