Fix Login modal on the AboutYou page in Checkout (#2817)
This fix is done by copying the solution from the default Login modal. This is not a solution to perpetuate moving forward as we have dangerous duplicated code. This solution is fixing a current production issue and should be seen as an intermediate step for when the Login modal is unified between Header and AboutYou.