TG-2309 [PayPal][FE] Add/update e2e tests for signup journey (#2814)
* [TG-2309] improve stability of signup-with-CO negative test

After 'TG-2237 [Prod] Apply promo code by clicking Get started' got
productionized, the e2e test was failing on the scenario of creating the new
user when there's no promo code conflict: the 'waitForElementVisible' creates
an assertion, which failed the test when the banner was not appearing.

The upshot of the failure was that every day, the first e2e run on staging was
failing, and restarting the e2e tests got the tests to pass.

The downside of the fix is that it increases the tests runtime by the default
interval (30s) even in cases where the signup request succeeds faster and leads
to appearance or non-appearance of the banner.  I found no way to catch the
request end reliably without waiting for the full timeout.

The unreliability in this test still persists in another scenario (failure on
checkIfErrorForCardDetailsVisible), but it occurs less frequently and more
randomly, so finding its cause remains to be done.

* [TG-2309] rename signup tests to include checkout.com

In preparation for PayPal signup flows.

* [TG-2309] merge two checkout pages into one

A page object should provide tools for working with a given section of the
site; it doesn't make sense to have a separate "error" object for the same
section, because it has duplicated code: increased clarity of the
error-scenario tools in a separate file doesn't outweigh the cost of having the
same functionality repeated in two files.

* [TG-2309] apply the same stability fix in the success-flow test

* [TG-2309] poll until condition instead of waiting full timeout

The synchronous logic of checkConditions would be:
```
for (const condition of conditions) {
  const { tag, checkFn } = condition
  const value = checkFn()
  if (value) {
    passedTag = tag
    break
  }
}

if (passedTag) {
  this.callback('conditionMet', passedTag)
} else {
  this.scheduleTimeout()
}
```

Since checkFn-s actually use continuation passing, a simple loop becomes
three functions.

* [TG-2309] reorganize tests in order to extract common parts

* [TG-2309] extract common parts of the flow into a single module

* [TG-2309] add utils for the PayPal flow

* [TG-2309] success scenario for signing up with PayPal

* [TG-2309] on mobile, make sure the consent button is clicked

* [TG-2309] fix missing zero that got lost while moving the constant

* [TG-2309] make small delay a constant; increase 1000 to 2000

Though the impetus for the change is refactoring, this is a functional change,
because I don't see the reason why 1 or 2 seconds should matter in cases where
the artificial delay is needed at all; and higher number is safer.

* [TG-2309] reduce duplication of hard-coded selectors

* [TG-2309] replace tabs with spaces

* [TG-2309] for login tests, use default 30s timeout instead of 15s

The login request can be slow at times, thus failing the tests.

* [TG-2309] improve reliability of logged-in-make-order tests

Sometimes the url check kicked in while the 'make-order' request was still in
progress (and the spinner on the Checkout button was working still on the Menu
page).  This change makes it so that the test waits for the full timeout before
failing.

Co-authored-by: Alexander Rymasheusky <alexander.rymashevskiy@gousto.co.uk>