[TG-2499] give some more time to paypal initialization (#2837)
The communication with PayPal appears to be slower on staging than on the local
env, hence the consent button didn't appear in 30s (default timeout), but the
screenshot made a bit later showed the window with the consent button.  Hence
the incresed timeout might help with that (and the process on the PayPal's side
after login and before the choose-card-and-consent screen shows up appears to
be quite involved indeed, with a lot of requests and redirects).