Fix usage of CSS variables (#2840)
Spacing should not be loaded through the sass-loader anymore as it might have unexpected effects (e.g. adding CSS at the end of the file that might override existing styles causing unintended effects). Also, the typography file was imported through the loader as well, but it was safe to remove as the variable was not used.