[TC-2003 Update cypress] Update version of Cypress to make it work locally again. (#3016)
Something must have changed in the Gousto web application, independently from gousto2webclient code. Cypress used to work locally a week ago, now it doesn't even going back in time with Git. An hypothesis is that something has changed that doesn't have to do with the code (A GTM tag for example?). What happens is that the page doesn't load, so `cy.visit('/')` never finishes. Even if the page looks loaded in the UI, the next command never runs.

Updating Cypress to a newer version worked, I guess this newer version is more robust and had this specific case in mind.

They changed `blacklist` and `whitelist` config keywords to `blocklist` and `ignore`