[TG-2778] Fetch new subscription endpoint (#3389)
* [TG-2778] Add /subscriptions to routes

* [TG-2778] Update buildSubscriptionCommandUrl and buildSubscriptionQueryUrl

* [TG-2778] Add isNewSubscriptionApiEnabled to Subscription container

Also creates a controller to manage whether subscription endpoint can be called in cases that it needs userId

* [TG-2778] Implement new subscription endpoint

It uses the new endpoint when isNewSubscriptionApiEnabled is true, then it'll parse the response to the current format that the Subscription component already uses.

* [TG-2778] Remove unused sku and price attributes

* [TG-2778] Add regression test coverage for the new subscription endpoint

* [TG-2778] Add fetchSubscriptionV2 method to api

* [TG-2778] Remove subscriptionLoadData from action::subscriptionPause

This same call was duplicate on the Account compoenent. It's already being called on userLoadData#L298.

* [TG-2778] Move subscription payload mapping function into a helper

As it's going to be reused on other actions.

* [TG-2778] Update useSubscriptionData to use the mapping helper

* [TG-2778] Add isNewSubscriptionApiEnabled to decide which fixture the regression test will wait for

* [TG-2778] Fix deactivateSubscriptionV2 URL

* [TG-2778] Fix subscription payload as it comes back with an extra `data` from our custom `fetch`

* [TG-2778] Small refactor on actions::user

This is to avoid - when on the experiment for the new subs. endpoint - the projected deliveries endpoint to be called without userId in the URL

* [TG-2778][TG-3893] Fix skip and unskip calls

Without the content/json the endpoint doesn't take the body request

* [TG-2778] Fix MyDeliveriesNew and Subscription regression tests

There was an issue with the way subscription payload was parsed so the Pause Subscription modal wasn't appearing on MyDeliveries as subscription.status was undefined.

* [TG-2778] Split subscriptionStatus in two tests

Due to issues with unstable feature flag, we've moved the experiment file into its own and set the flag before going to the subs page.