Added Dockerfile that creates an image of the service
This image now actually has the webclient service and runs it, rather than just acting as a place to build Webclient bind mounted inside from the host.

Added docker-compose file that joins an external network called dps that is created by the G2FE docker-compose file.

Added a dev version of this file that bind mounts the local repo in to the container for development purposes.

Below is a guide to the run script commands that have been added.

## run
Runs the service in docker, connected to the `dps` network and also bound to port 8080 on the host. More on this below.

## dev --docker
Runs Webclient in docker, with bind mounts for local development.

If doing this, you'll have to run `docker compose exec webclient bash` to get in to the running container or possibly even `run webclient bash` would work. Then from there `cd` run `npm run watch` to compile changes. If you are running in docker, running `npm run watch` outside of docker will fail unless you run `npm rebuild node-sass` as when creating the image, `node-sass` will have been compiled for the *nix architecture of the running container (linux) rather than Mac OS.

When running in docker, the service both joins the `dps` network and also binds to port 8080 on the host, so you can always go to `frontend.gousto.local:8080` to get access to Webclient directly, bypassing G2FE and Docker containers can access it via `webclient:8080` if they are on the `dps` network. This allows various run methods of G2FE to work, just in case I don't merge the branch that has specific commands for pointing G2FE to Webclient running within Docker and outside of Docker (which due to the added complexity I might not).

## dev --host
Alternatively the above command allows you to run the entire service on the host.

This command runs npm run watch and puts it in the background and then runs npm run start:dev.

Since npm run watch is a continuous process that never exits, a single & has had to be used to run the two processes in parallel. This created a problem where Nodemon would try to watch for dist/server.js before it had been created and crash without restarting on file change. For this reason I have added a touch command to start:dev to create the watched file before it starts Nodemon. This does not overwrite already existing files by the way, for example when starting up for a second time. The build watcher then kicks in and compiles the real server, the change of which Nodemon then picks up and restarts the server.

A way to run G2FE pointing at Webclient running locally has been made. See the branch in G2FE with the corresponding name as this one.

Updated the readme.md to reflect new changes and the new run script.