[TC-2524] Fix UI issue with modals appearing under the menu on small screens (#3472)
On smaller screens, if the side header menu is open, it would appear OVER any modal, so, in the scenario of being logged out, the login modal would appear under the menu. This fix pushes the modal to effectively be on top of other elements, as it should be (modals normally take precedence as they act as takeovers of the page).