[FEF-21] - [HOTFIX] -Added max-http-header-size option to all Cypress calls (#3483)
For some reason after upgrading to Cypress 7, certain tests were failing as mocks were returning HTTP status code `431 Request Header Fields Too Large`, but only when pointing to staging.

This has happened previously with Cypress as shown [here](https://github.com/cypress-io/cypress/issues/5431) but I think this is largely unrelated.

Upon analysis of the headers, they are very large and Cypress throws a load of extra properties on to them. Given that our cookie header is large and couple that with the addition of the Cypress ones, it seems the difference when running locally between the number of characters in `frontend.gousto.local` and `staging-frontend.gousto.info` and a few extra redirect properties were enough to push it over the Node JS mandated 8KB.

Either way, this property fixes the problem, however ugly it may be. We should probably look at the size of our cookies.