[TG-3698] Add reCaptcha to e-mail on Refer A Friend (#3449)
* [TG-3698] Isolate a shallow wrapper version for each describe

It will help on write tests with different versions of the wrap. E.g. feature X flag on

* [TG-3698] Implement captcha on <ReferAFriend>

Google reCAPTCHA added to when an user tries to refer a friend via e-mail.

The currently captcha functionality is copied from the <LoginForm>.

* [TG-3698] Move isSecretPingdomEmail logic into a helper

As now both ReferAFriend and LoginForm share that logic.

* [TG-3698] Fix eslint error about deprecated methods (componentWillMount, componentWillReceiveProps) on LoginForm

* [TG-3698] Call webclient refer a friend endpoint

* [TG-3698] Fetch features from S3

So we can use isRecaptchaEnabled

* [TG-3698] Get reCaptcha secrets to S3

* [TG-3698] Add more test coverage

* [TG-3698] CR | Add RAF recapcha to custom env vars

And default it to the test key. Also removes the logic that handles the keys based on the __ENV__ value as node-config files should already be doing that work.

* [TG-3698] Add tests for recaptcha on Login

Also makes a small change to how the recaptcha update `ref` by extrating it to a separate function so it can be tested.

* [TG-3698] Add RAF public key to be attached to the client config

* Use correct method for sending referral request to server

* Add correct callback for captcha

* Linting fix

* Restore of removed during merge referAfriend method

* Improve test coverage after merge

Co-authored-by: Kiryl Surahatau <kirill.surogatov@gousto.co.uk>