[TG-2257] Remove deprecated React lifecycle methods (#3494)
* [TG-2257] fix proptype warning

Even though it doesn't strictly relate to react methods, it's a tiny change
that doesn't warrant a separate ticket.

* [TG-2257] remove componentWillReceiveProps from Signup.js

Since changeStep() is invoked from elsewhere on navigating to wizard and
between its steps, the logic in this method was redundant.

* [TG-2257] replace willMount to didMount in PromoCode

There's no difference in observable behavior.

* [TG-2257] replace willMount with didMount in Address

There is no observable difference in behavior.

* [TG-2257] replace willReceiveProps with didUpdate in Delivery.js

Couldn't reproduce the scrolling behavior (probably it's become irrelevant
after checkout overhaul); but to be safe retained it; since it's an animation
didUpdate fits per the conversion guidelines.

* [TG-2257] replace willReceiveProps with didUpdate in DeliveryInstruction

No observable difference in behavior.

* [TG-2257] remove willReceiveProps from Checkout

The method checked tariffId which is a dead concept as confirmed by Carrots.

* [TG-2257] remove autoFocus-related props of TextInput

As autoFocus was removed without explanation in 6d1313a2f, the prop and
willReceiveProps logic became obsolete at that point.  It makes no sense to
keep it longer.

* [TG-2257] replace willReceiveProps with didUpdate in ScrollToRefs

Couldn't reproduce the scrolling behavior (it's used only in Checkout forms;
probably because irrelevant after checkout overhaul); but retained it just in
case.