Reverted recent change to OrderedRecipe that passes a url string to GoustoImage.
If the recipe images aren't in the store then undefined is passed as the media string to the GoustoImage component. This breaks the prop-types and the component itself which expects it to be either a string or an Immutable list.

It falls down at 87 `media.size` which is reached if the media prop isn't a string. `undefined` of course, does not have a size property, so this causes an exception to be thrown.

I'm not sure why this was changed in the first place to be honest, so I've done a commit to revert it back to what it should be (given the prop types of GoustoImage).